NATIVE-42: Move Events component to components folder
