+++ src/routes/events/components/Events.js
@@ -7,7 +7,7 @@ import { EventModel } from '@integreat-app/integreat-api-client'
 import Page from '../../../modules/common/components/Page'
 import ContentNotFoundError from '../../../modules/common/errors/ContentNotFoundError'
 import PageDetail from '../../../modules/common/components/PageDetail'
-import EventListItem from '../components/EventListItem'
+import EventListItem from './EventListItem'
 import List from '../../../modules/common/components/List'
 import Caption from '../../../modules/common/components/Caption'
 import Failure from '../../error/components/Failure'
